refactor: add BaseSchemaComponent
